Removing use of ssize_t for windows
git-svn-id: https://svn.code.sf.net/p/stldb/code/branches/r1.2@434 76cca37e-3e0e-4c71-b5a1-d872068be9f0

+++ stldb_lib/stldb/checkpoint.h
@@ -347,7 +347,7 @@ private:
 		// TODO - should this be in network byte order for portability.?
 		// are binary streams from Boost.Serialization portable to some extent?
 		_checkpoint.filestream.read(reinterpret_cast<char*>(&size), sizeof(size));
-		BOOST_ASSERT(i == _checkpoint.free_by_offset.end() || _offset + (ssize_t)size <= i->first);
+		BOOST_ASSERT(i == _checkpoint.free_by_offset.end() || _offset + size <= i->first);
 		boost_iarchive_t archive(_checkpoint.filestream, boost::archive::no_header);
         archive & _current;
 		if (!_checkpoint.filestream) {
